[Draft] Split Dimension object and reorganize of annotation objects
part 1
