+++ src/Mod/Draft/Draft.py
@@ -251,79 +251,14 @@ def makeRectangle(length, height, placement=None, face=None, support=None):
 
     return obj
 
-def makeDimension(p1,p2,p3=None,p4=None):
-    """makeDimension(p1,p2,[p3]) or makeDimension(object,i1,i2,p3)
-    or makeDimension(objlist,indices,p3): Creates a Dimension object with
-    the dimension line passign through p3.The current line width and color
-    will be used. There are multiple  ways to create a dimension, depending on
-    the arguments you pass to it:
-    - (p1,p2,p3): creates a standard dimension from p1 to p2
-    - (object,i1,i2,p3): creates a linked dimension to the given object,
-    measuring the distance between its vertices indexed i1 and i2
-    - (object,i1,mode,p3): creates a linked dimension
-    to the given object, i1 is the index of the (curved) edge to measure,
-    and mode is either "radius" or "diameter".
-    """
-    if not FreeCAD.ActiveDocument:
-        FreeCAD.Console.PrintError("No active document. Aborting\n")
-        return
-    obj = FreeCAD.ActiveDocument.addObject("App::FeaturePython","Dimension")
-    _Dimension(obj)
-    if gui:
-        _ViewProviderDimension(obj.ViewObject)
-    if isinstance(p1,Vector) and isinstance(p2,Vector):
-        obj.Start = p1
-        obj.End = p2
-        if not p3:
-            p3 = p2.sub(p1)
-            p3.multiply(0.5)
-            p3 = p1.add(p3)
-    elif isinstance(p2,int) and isinstance(p3,int):
-        l = []
-        idx = (p2,p3)
-        l.append((p1,"Vertex"+str(p2+1)))
-        l.append((p1,"Vertex"+str(p3+1)))
-        obj.LinkedGeometry = l
-        obj.Support = p1
-        p3 = p4
-        if not p3:
-            v1 = obj.Base.Shape.Vertexes[idx[0]].Point
-            v2 = obj.Base.Shape.Vertexes[idx[1]].Point
-            p3 = v2.sub(v1)
-            p3.multiply(0.5)
-            p3 = v1.add(p3)
-    elif isinstance(p3,str):
-        l = []
-        l.append((p1,"Edge"+str(p2+1)))
-        if p3 == "radius":
-            #l.append((p1,"Center"))
-            if FreeCAD.GuiUp:
-                obj.ViewObject.Override = "R $dim"
-            obj.Diameter = False
-        elif p3 == "diameter":
-            #l.append((p1,"Diameter"))
-            if FreeCAD.GuiUp:
-                obj.ViewObject.Override = "Ø $dim"
-            obj.Diameter = True
-        obj.LinkedGeometry = l
-        obj.Support = p1
-        p3 = p4
-        if not p3:
-            p3 = p1.Shape.Edges[p2].Curve.Center.add(Vector(1,0,0))
-    obj.Dimline = p3
-    if hasattr(FreeCAD,"DraftWorkingPlane"):
-        normal = FreeCAD.DraftWorkingPlane.axis
-    else:
-        normal = FreeCAD.Vector(0,0,1)
-    if gui:
-        # invert the normal if we are viewing it from the back
-        vnorm = get3DView().getViewDirection()
-        if vnorm.getAngle(normal) < math.pi/2:
-            normal = normal.negative()
-    obj.Normal = normal
-    if gui:
-        formatObject(obj)
-        select(obj)
+from draftobjects.dimension import make_dimension
+makeDimension = make_dimension
+
+from draftobjects.dimension import LinearDimension
+_Dimension = LinearDimension
+
+from draftviewproviders.view_dimension import ViewProviderLinearDimension
+_ViewProviderDimension = ViewProviderLinearDimension
 
     return obj
 
@@ -3367,722 +3302,6 @@ class _ViewProviderDraftLink:
             return [obj.Base]
         else:
             return obj.ElementList
-
-class _Dimension(_DraftObject):
-    """The Draft Dimension object"""
-    def __init__(self, obj):
-        _DraftObject.__init__(self,obj,"Dimension")
-        
-        # Annotation
-        obj.addProperty("App::PropertyLink","DimensionStyle",
-                        "Annotation",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "Link dimension style"))
-
-        # Draft
-        obj.addProperty("App::PropertyVectorDistance","Start",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "Startpoint of dimension"))
-
-        obj.addProperty("App::PropertyVectorDistance","End",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "Endpoint of dimension"))
-
-        obj.addProperty("App::PropertyVector","Normal",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "The normal direction of this dimension"))
-
-        obj.addProperty("App::PropertyVector","Direction",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "The normal direction of this dimension"))
-
-        obj.addProperty("App::PropertyVectorDistance","Dimline",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "Point through which the dimension line passes"))
-
-        obj.addProperty("App::PropertyLink","Support",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "The object measured by this dimension"))
-
-        obj.addProperty("App::PropertyLinkSubList","LinkedGeometry",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "The geometry this dimension is linked to"))
-
-        obj.addProperty("App::PropertyLength","Distance",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "The measurement of this dimension"))
-
-        obj.addProperty("App::PropertyBool","Diameter",
-                        "Draft",
-                        QT_TRANSLATE_NOOP("App::Property",
-                                          "For arc/circle measurements, false = radius, true = diameter"))
-        obj.Start = FreeCAD.Vector(0,0,0)
-        obj.End = FreeCAD.Vector(1,0,0)
-        obj.Dimline = FreeCAD.Vector(0,1,0)
-        obj.Normal = FreeCAD.Vector(0,0,1)
-
-    def onChanged(self,obj,prop):
-        if hasattr(obj, "Distance"):
-            obj.setEditorMode('Distance', 1)
-        #if hasattr(obj,"Normal"):
-        #    obj.setEditorMode('Normal', 2)
-        if hasattr(obj, "Support"):
-            obj.setEditorMode('Support', 2)
-        if prop == "DimensionStyle":
-            if hasattr(obj, "DimensionStyle"):
-                from draftutils import gui_utils
-                gui_utils.format_object(target = obj, origin = obj.DimensionStyle)
-
-
-    def execute(self, obj):
-        import DraftGeomUtils
-        # set start point and end point according to the linked geometry
-        if obj.LinkedGeometry:
-            if len(obj.LinkedGeometry) == 1:
-                lobj = obj.LinkedGeometry[0][0]
-                lsub = obj.LinkedGeometry[0][1]
-                if len(lsub) == 1:
-                    if "Edge" in lsub[0]:
-                        n = int(lsub[0][4:])-1
-                        edge = lobj.Shape.Edges[n]
-                        if DraftGeomUtils.geomType(edge) == "Line":
-                            obj.Start = edge.Vertexes[0].Point
-                            obj.End = edge.Vertexes[-1].Point
-                        elif DraftGeomUtils.geomType(edge) == "Circle":
-                            c = edge.Curve.Center
-                            r = edge.Curve.Radius
-                            a = edge.Curve.Axis
-                            ray = obj.Dimline.sub(c).projectToPlane(Vector(0,0,0),a)
-                            if (ray.Length == 0):
-                                ray = a.cross(Vector(1,0,0))
-                                if (ray.Length == 0):
-                                    ray = a.cross(Vector(0,1,0))
-                            ray = DraftVecUtils.scaleTo(ray,r)
-                            if hasattr(obj,"Diameter"):
-                                if obj.Diameter:
-                                    obj.Start = c.add(ray.negative())
-                                    obj.End = c.add(ray)
-                                else:
-                                    obj.Start = c
-                                    obj.End = c.add(ray)
-                elif len(lsub) == 2:
-                    if ("Vertex" in lsub[0]) and ("Vertex" in lsub[1]):
-                        n1 = int(lsub[0][6:])-1
-                        n2 = int(lsub[1][6:])-1
-                        obj.Start = lobj.Shape.Vertexes[n1].Point
-                        obj.End = lobj.Shape.Vertexes[n2].Point
-            elif len(obj.LinkedGeometry) == 2:
-                lobj1 = obj.LinkedGeometry[0][0]
-                lobj2 = obj.LinkedGeometry[1][0]
-                lsub1 = obj.LinkedGeometry[0][1]
-                lsub2 = obj.LinkedGeometry[1][1]
-                if (len(lsub1) == 1) and (len(lsub2) == 1):
-                    if ("Vertex" in lsub1[0]) and ("Vertex" in lsub2[1]):
-                        n1 = int(lsub1[0][6:])-1
-                        n2 = int(lsub2[0][6:])-1
-                        obj.Start = lobj1.Shape.Vertexes[n1].Point
-                        obj.End = lobj2.Shape.Vertexes[n2].Point
-        # set the distance property
-        total_len = (obj.Start.sub(obj.End)).Length
-        if round(obj.Distance.Value, precision()) != round(total_len, precision()):
-            obj.Distance = total_len
-        if gui:
-            if obj.ViewObject:
-                obj.ViewObject.update()
-
-
-class _ViewProviderDimension(_ViewProviderDraft):
-    """
-    A View Provider for the Draft Dimension object
-    
-    DIMENSION VIEW PROVIDER:
-    
-        |              txt               |       e
-    ----o--------------------------------o-----
-        |                                |
-        |                                |       d  
-        |                                |
-
-     a  b               c                b  a
-    
-    a = DimOvershoot (vobj)
-    b = Arrows (vobj)
-    c = Dimline (obj)
-    d = ExtLines (vobj)
-    e = ExtOvershoot (vobj)
-    txt = label (vobj)
-
-    STRUCTURE:
-    vobj.node.color
-             .drawstyle
-             .lineswitch1.coords
-                         .line
-                         .marks
-                         .marksDimOvershoot
-                         .marksExtOvershoot
-             .label.textpos
-                   .color
-                   .font
-                   .text
-             
-    vobj.node3d.color
-               .drawstyle
-               .lineswitch3.coords
-                           .line
-                           .marks
-                           .marksDimOvershoot
-                           .marksExtOvershoot
-               .label3d.textpos
-                       .color
-                       .font3d
-                       .text3d
-    
-    """
-    def __init__(self, obj): 
-        # annotation properties
-        obj.addProperty("App::PropertyFloat","ScaleMultiplier",
-                        "Annotation",QT_TRANSLATE_NOOP("App::Property",
-                        "Dimension size overall multiplier"))
-
-        # text properties
-        obj.addProperty("App::PropertyFont","FontName",
-                        "Text",QT_TRANSLATE_NOOP("App::Property","Font name"))
-        obj.addProperty("App::PropertyLength","FontSize",
-                        "Text",QT_TRANSLATE_NOOP("App::Property","Font size"))
-        obj.addProperty("App::PropertyLength","TextSpacing",
-                        "Text",QT_TRANSLATE_NOOP("App::Property",
-                        "The spacing between the text and the dimension line"))
-        obj.addProperty("App::PropertyBool","FlipText",
-                        "Text",QT_TRANSLATE_NOOP("App::Property",
-                        "Rotate the dimension text 180 degrees"))
-        obj.addProperty("App::PropertyVectorDistance","TextPosition",
-                        "Text",QT_TRANSLATE_NOOP("App::Property",
-                        "The position of the text. Leave (0,0,0) for automatic position"))
-        obj.addProperty("App::PropertyString","Override",
-                        "Text",QT_TRANSLATE_NOOP("App::Property",
-                        "Text override. Use $dim to insert the dimension length"))
-        # units properties
-        obj.addProperty("App::PropertyInteger","Decimals",
-                        "Units",QT_TRANSLATE_NOOP("App::Property",
-                        "The number of decimals to show"))
-        obj.addProperty("App::PropertyBool","ShowUnit",
-                        "Units",QT_TRANSLATE_NOOP("App::Property",
-                        "Show the unit suffix"))
-        obj.addProperty("App::PropertyString","UnitOverride",
-                        "Units",QT_TRANSLATE_NOOP("App::Property",
-                        "A unit to express the measurement. Leave blank for system default"))
-        # graphics properties
-        obj.addProperty("App::PropertyFloat","LineWidth",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property","Line width"))
-        obj.addProperty("App::PropertyColor","LineColor",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property","Line color"))                
-        obj.addProperty("App::PropertyLength","ArrowSize",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property","Arrow size"))
-        obj.addProperty("App::PropertyEnumeration","ArrowType",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property","Arrow type"))
-        obj.addProperty("App::PropertyBool","FlipArrows",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property",
-                        "Rotate the dimension arrows 180 degrees"))        
-        obj.addProperty("App::PropertyDistance","DimOvershoot",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property",
-                        "The distance the dimension line is extended past the extension lines"))        
-        obj.addProperty("App::PropertyDistance","ExtLines",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property",
-                        "Length of the extension lines"))
-        obj.addProperty("App::PropertyDistance","ExtOvershoot",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property",
-                        "Length of the extension line above the dimension line"))
-        obj.addProperty("App::PropertyBool","ShowLine",
-                        "Graphics",QT_TRANSLATE_NOOP("App::Property",
-                        "Shows the dimension line and arrows"))
-        
-        param = FreeCAD.ParamGet("User parameter:BaseApp/Preferences/Mod/Draft")
-        annotation_scale = param.GetFloat("DraftAnnotationScale", 1.0)
-        obj.ScaleMultiplier = 1 / annotation_scale
-        obj.FontSize = getParam("textheight",0.20)
-        obj.TextSpacing = getParam("dimspacing",0.05)
-        obj.FontName = getParam("textfont","")
-        obj.ArrowSize = getParam("arrowsize",0.1)
-        obj.ArrowType = arrowtypes
-        obj.ArrowType = arrowtypes[getParam("dimsymbol",0)]
-        obj.ExtLines = getParam("extlines",0.3)
-        obj.DimOvershoot = getParam("dimovershoot",0)
-        obj.ExtOvershoot = getParam("extovershoot",0)
-        obj.Decimals = getParam("dimPrecision",2)
-        obj.ShowUnit = getParam("showUnit",True)
-        obj.ShowLine = True
-        _ViewProviderDraft.__init__(self,obj)
-
-    def attach(self, vobj):
-        """called on object creation"""
-        from pivy import coin
-        self.Object = vobj.Object
-        self.color = coin.SoBaseColor()
-        self.font = coin.SoFont()
-        self.font3d = coin.SoFont()
-        self.text = coin.SoAsciiText()
-        self.text3d = coin.SoText2()
-        self.text.string = "d" # some versions of coin crash if string is not set
-        self.text3d.string = "d"
-        self.textpos = coin.SoTransform()
-        self.text.justification = self.text3d.justification = coin.SoAsciiText.CENTER
-        label = coin.SoSeparator()
-        label.addChild(self.textpos)
-        label.addChild(self.color)
-        label.addChild(self.font)
-        label.addChild(self.text)
-        label3d = coin.SoSeparator()
-        label3d.addChild(self.textpos)
-        label3d.addChild(self.color)
-        label3d.addChild(self.font3d)
-        label3d.addChild(self.text3d)
-        self.coord1 = coin.SoCoordinate3()
-        self.trans1 = coin.SoTransform()
-        self.coord2 = coin.SoCoordinate3()
-        self.trans2 = coin.SoTransform()
-        self.transDimOvershoot1 = coin.SoTransform()
-        self.transDimOvershoot2 = coin.SoTransform()
-        self.transExtOvershoot1 = coin.SoTransform()
-        self.transExtOvershoot2 = coin.SoTransform()
-        self.marks = coin.SoSeparator()
-        self.marksDimOvershoot = coin.SoSeparator()
-        self.marksExtOvershoot = coin.SoSeparator()
-        self.drawstyle = coin.SoDrawStyle()
-        self.line = coin.SoType.fromName("SoBrepEdgeSet").createInstance()
-        self.coords = coin.SoCoordinate3()
-        self.node = coin.SoGroup()
-        self.node.addChild(self.color)
-        self.node.addChild(self.drawstyle)
-        self.lineswitch2 = coin.SoSwitch()
-        self.lineswitch2.whichChild = -3
-        self.node.addChild(self.lineswitch2)
-        self.lineswitch2.addChild(self.coords)
-        self.lineswitch2.addChild(self.line)
-        self.lineswitch2.addChild(self.marks)
-        self.lineswitch2.addChild(self.marksDimOvershoot)
-        self.lineswitch2.addChild(self.marksExtOvershoot)
-        self.node.addChild(label)
-        self.node3d = coin.SoGroup()
-        self.node3d.addChild(self.color)
-        self.node3d.addChild(self.drawstyle)
-        self.lineswitch3 = coin.SoSwitch()
-        self.lineswitch3.whichChild = -3
-        self.node3d.addChild(self.lineswitch3)
-        self.lineswitch3.addChild(self.coords)
-        self.lineswitch3.addChild(self.line)
-        self.lineswitch3.addChild(self.marks)
-        self.lineswitch3.addChild(self.marksDimOvershoot)
-        self.lineswitch3.addChild(self.marksExtOvershoot)
-        self.node3d.addChild(label3d)
-        vobj.addDisplayMode(self.node,"2D")
-        vobj.addDisplayMode(self.node3d,"3D")
-        self.updateData(vobj.Object,"Start")
-        self.onChanged(vobj,"FontSize")
-        self.onChanged(vobj,"FontName")
-        self.onChanged(vobj,"ArrowType")
-        self.onChanged(vobj,"LineColor")
-        self.onChanged(vobj,"DimOvershoot")
-        self.onChanged(vobj,"ExtOvershoot")
-
-    def updateData(self, obj, prop):
-        """called when the base object is changed"""
-        import DraftGui
-        if prop in ["Start","End","Dimline","Direction"]:
-
-            if obj.Start == obj.End:
-                return
-
-            if not hasattr(self,"node"):
-                return
-
-            import Part, DraftGeomUtils
-            from pivy import coin
-
-            # calculate the 4 points
-            self.p1 = obj.Start
-            self.p4 = obj.End
-            base = None
-            if hasattr(obj,"Direction"):
-                if not DraftVecUtils.isNull(obj.Direction):
-                    v2 = self.p1.sub(obj.Dimline)
-                    v3 = self.p4.sub(obj.Dimline)
-                    v2 = DraftVecUtils.project(v2,obj.Direction)
-                    v3 = DraftVecUtils.project(v3,obj.Direction)
-                    self.p2 = obj.Dimline.add(v2)
-                    self.p3 = obj.Dimline.add(v3)
-                    if DraftVecUtils.equals(self.p2,self.p3):
-                        base = None
-                        proj = None
-                    else:
-                        base = Part.LineSegment(self.p2,self.p3).toShape()
-                        proj = DraftGeomUtils.findDistance(self.p1,base)
-                        if proj:
-                            proj = proj.negative()
-            if not base:
-                if DraftVecUtils.equals(self.p1,self.p4):
-                    base = None
-                    proj = None
-                else:
-                    base = Part.LineSegment(self.p1,self.p4).toShape()
-                    proj = DraftGeomUtils.findDistance(obj.Dimline,base)
-                if proj:
-                    self.p2 = self.p1.add(proj.negative())
-                    self.p3 = self.p4.add(proj.negative())
-                else:
-                    self.p2 = self.p1
-                    self.p3 = self.p4
-            if proj:
-                if hasattr(obj.ViewObject,"ExtLines") and hasattr(obj.ViewObject,"ScaleMultiplier"):
-                    dmax = obj.ViewObject.ExtLines.Value * obj.ViewObject.ScaleMultiplier
-                    if dmax and (proj.Length > dmax):
-                        if (dmax > 0):
-                            self.p1 = self.p2.add(DraftVecUtils.scaleTo(proj,dmax))
-                            self.p4 = self.p3.add(DraftVecUtils.scaleTo(proj,dmax))
-                        else:
-                            rest = proj.Length + dmax
-                            self.p1 = self.p2.add(DraftVecUtils.scaleTo(proj,rest))
-                            self.p4 = self.p3.add(DraftVecUtils.scaleTo(proj,rest))
-            else:
-                proj = (self.p3.sub(self.p2)).cross(Vector(0,0,1))
-
-            # calculate the arrows positions
-            self.trans1.translation.setValue((self.p2.x,self.p2.y,self.p2.z))
-            self.coord1.point.setValue((self.p2.x,self.p2.y,self.p2.z))
-            self.trans2.translation.setValue((self.p3.x,self.p3.y,self.p3.z))
-            self.coord2.point.setValue((self.p3.x,self.p3.y,self.p3.z))
-
-            # calculate dimension and extension lines overshoots positions
-            self.transDimOvershoot1.translation.setValue((self.p2.x,self.p2.y,self.p2.z))
-            self.transDimOvershoot2.translation.setValue((self.p3.x,self.p3.y,self.p3.z))
-            self.transExtOvershoot1.translation.setValue((self.p2.x,self.p2.y,self.p2.z))
-            self.transExtOvershoot2.translation.setValue((self.p3.x,self.p3.y,self.p3.z))
-
-            # calculate the text position and orientation
-            if hasattr(obj,"Normal"):
-                if DraftVecUtils.isNull(obj.Normal):
-                    if proj:
-                        norm = (self.p3.sub(self.p2).cross(proj)).negative()
-                    else:
-                        norm = Vector(0,0,1)
-                else:
-                    norm = FreeCAD.Vector(obj.Normal)
-            else:
-                if proj:
-                    norm = (self.p3.sub(self.p2).cross(proj)).negative()
-                else:
-                    norm = Vector(0,0,1)
-            if not DraftVecUtils.isNull(norm):
-                norm.normalize()
-            u = self.p3.sub(self.p2)
-            u.normalize()
-            v1 = norm.cross(u)
-            rot1 = FreeCAD.Placement(DraftVecUtils.getPlaneRotation(u,v1,norm)).Rotation.Q
-            self.transDimOvershoot1.rotation.setValue((rot1[0],rot1[1],rot1[2],rot1[3]))
-            self.transDimOvershoot2.rotation.setValue((rot1[0],rot1[1],rot1[2],rot1[3]))
-            if hasattr(obj.ViewObject,"FlipArrows"):
-                if obj.ViewObject.FlipArrows:
-                    u = u.negative()
-            v2 = norm.cross(u)
-            rot2 = FreeCAD.Placement(DraftVecUtils.getPlaneRotation(u,v2,norm)).Rotation.Q
-            self.trans1.rotation.setValue((rot2[0],rot2[1],rot2[2],rot2[3]))
-            self.trans2.rotation.setValue((rot2[0],rot2[1],rot2[2],rot2[3]))
-            if self.p1 != self.p2:
-                u3 = self.p1.sub(self.p2)
-                u3.normalize()
-                v3 = norm.cross(u3)
-                rot3 = FreeCAD.Placement(DraftVecUtils.getPlaneRotation(u3,v3,norm)).Rotation.Q
-                self.transExtOvershoot1.rotation.setValue((rot3[0],rot3[1],rot3[2],rot3[3]))
-                self.transExtOvershoot2.rotation.setValue((rot3[0],rot3[1],rot3[2],rot3[3]))
-            if hasattr(obj.ViewObject,"TextSpacing") and hasattr(obj.ViewObject,"ScaleMultiplier"):
-                ts = obj.ViewObject.TextSpacing.Value * obj.ViewObject.ScaleMultiplier
-                offset = DraftVecUtils.scaleTo(v1,ts)
-            else:
-                offset = DraftVecUtils.scaleTo(v1,0.05)
-            rott = rot1
-            if hasattr(obj.ViewObject,"FlipText"):
-                if obj.ViewObject.FlipText:
-                    rott = FreeCAD.Rotation(*rott).multiply(FreeCAD.Rotation(norm,180)).Q
-                    offset = offset.negative()
-            # setting text
-            try:
-                m = obj.ViewObject.DisplayMode
-            except: # swallow all exceptions here since it always fails on first run (Displaymode enum no set yet)
-                m = ["2D","3D"][getParam("dimstyle",0)]
-            if m == "3D":
-                offset = offset.negative()
-            self.tbase = (self.p2.add((self.p3.sub(self.p2).multiply(0.5)))).add(offset)
-            if hasattr(obj.ViewObject,"TextPosition"):
-                if not DraftVecUtils.isNull(obj.ViewObject.TextPosition):
-                    self.tbase = obj.ViewObject.TextPosition
-            self.textpos.translation.setValue([self.tbase.x,self.tbase.y,self.tbase.z])
-            self.textpos.rotation = coin.SbRotation(rott[0],rott[1],rott[2],rott[3])
-            su = True
-            if hasattr(obj.ViewObject,"ShowUnit"):
-                su = obj.ViewObject.ShowUnit
-            # set text value
-            l = self.p3.sub(self.p2).Length
-            unit = None
-            if hasattr(obj.ViewObject,"UnitOverride"):
-                unit = obj.ViewObject.UnitOverride
-            # special representation if "Building US" scheme
-            if FreeCAD.ParamGet("User parameter:BaseApp/Preferences/Units").GetInt("UserSchema",0) == 5:
-                s = FreeCAD.Units.Quantity(l,FreeCAD.Units.Length).UserString
-                self.string = s.replace("' ","'- ")
-                self.string = s.replace("+"," ")
-            elif hasattr(obj.ViewObject,"Decimals"):
-                self.string = DraftGui.displayExternal(l,obj.ViewObject.Decimals,'Length',su,unit)
-            else:
-                self.string = DraftGui.displayExternal(l,None,'Length',su,unit)
-            if hasattr(obj.ViewObject,"Override"):
-                if obj.ViewObject.Override:
-                    self.string = obj.ViewObject.Override.replace("$dim",\
-                            self.string)
-            self.text.string = self.text3d.string = stringencodecoin(self.string)
-
-            # set the lines
-            if m == "3D":
-                # calculate the spacing of the text
-                textsize = (len(self.string)*obj.ViewObject.FontSize.Value)/4.0
-                spacing = ((self.p3.sub(self.p2)).Length/2.0) - textsize
-                self.p2a = self.p2.add(DraftVecUtils.scaleTo(self.p3.sub(self.p2),spacing))
-                self.p2b = self.p3.add(DraftVecUtils.scaleTo(self.p2.sub(self.p3),spacing))
-                self.coords.point.setValues([[self.p1.x,self.p1.y,self.p1.z],
-                                             [self.p2.x,self.p2.y,self.p2.z],
-                                             [self.p2a.x,self.p2a.y,self.p2a.z],
-                                             [self.p2b.x,self.p2b.y,self.p2b.z],
-                                             [self.p3.x,self.p3.y,self.p3.z],
-                                             [self.p4.x,self.p4.y,self.p4.z]])
-                #self.line.numVertices.setValues([3,3])
-                self.line.coordIndex.setValues(0,7,(0,1,2,-1,3,4,5))
-            else:
-                self.coords.point.setValues([[self.p1.x,self.p1.y,self.p1.z],
-                                             [self.p2.x,self.p2.y,self.p2.z],
-                                             [self.p3.x,self.p3.y,self.p3.z],
-                                             [self.p4.x,self.p4.y,self.p4.z]])
-                #self.line.numVertices.setValue(4)
-                self.line.coordIndex.setValues(0,4,(0,1,2,3))
-
-    def onChanged(self, vobj, prop):
-        """called when a view property has changed"""
-        if prop == "ScaleMultiplier" and hasattr(vobj,"ScaleMultiplier"):
-            # update all dimension values
-            if hasattr(self,"font"):
-                self.font.size = vobj.FontSize.Value*vobj.ScaleMultiplier
-            if hasattr(self,"font3d"):
-                self.font3d.size = vobj.FontSize.Value*100*vobj.ScaleMultiplier
-            if hasattr(self,"node") and hasattr(self,"p2") and hasattr(vobj,"ArrowSize"):
-                self.remove_dim_arrows()
-                self.draw_dim_arrows(vobj)
-            if hasattr(vobj,"DimOvershoot"):
-                self.remove_dim_overshoot()
-                self.draw_dim_overshoot(vobj)
-            if hasattr(vobj,"ExtOvershoot"):
-                self.remove_ext_overshoot()
-                self.draw_ext_overshoot(vobj)
-            self.updateData(vobj.Object,"Start")
-            vobj.Object.touch()
-            
-        elif (prop == "FontSize") and hasattr(vobj,"FontSize"):
-            if hasattr(self,"font") and hasattr(vobj,"ScaleMultiplier"):
-                self.font.size = vobj.FontSize.Value*vobj.ScaleMultiplier
-            if hasattr(self,"font3d") and hasattr(vobj,"ScaleMultiplier"):
-                self.font3d.size = vobj.FontSize.Value*100*vobj.ScaleMultiplier
-            vobj.Object.touch()
-            
-        elif (prop == "FontName") and hasattr(vobj,"FontName"):
-            if hasattr(self,"font") and hasattr(self,"font3d"):
-                self.font.name = self.font3d.name = str(vobj.FontName)
-                vobj.Object.touch()
-                
-        elif (prop == "LineColor") and hasattr(vobj,"LineColor"):
-            if hasattr(self,"color"):
-                c = vobj.LineColor
-                self.color.rgb.setValue(c[0],c[1],c[2])
-                
-        elif (prop == "LineWidth") and hasattr(vobj,"LineWidth"):
-            if hasattr(self,"drawstyle"):
-                self.drawstyle.lineWidth = vobj.LineWidth
-                
-        elif (prop in ["ArrowSize","ArrowType"]) and hasattr(vobj,"ArrowSize"):
-            if hasattr(self,"node") and hasattr(self,"p2"):
-                if hasattr(vobj,"ScaleMultiplier"):
-                    self.remove_dim_arrows()
-                    self.draw_dim_arrows(vobj)
-                    vobj.Object.touch()
-                
-        elif (prop == "DimOvershoot") and hasattr(vobj,"DimOvershoot"):
-            if hasattr(vobj,"ScaleMultiplier"):
-                self.remove_dim_overshoot()
-                self.draw_dim_overshoot(vobj)
-                vobj.Object.touch()
-            
-        elif (prop == "ExtOvershoot") and hasattr(vobj,"ExtOvershoot"):
-            if hasattr(vobj,"ScaleMultiplier"):
-                self.remove_ext_overshoot()
-                self.draw_ext_overshoot(vobj)
-                vobj.Object.touch()
-            
-        elif (prop == "ShowLine") and hasattr(vobj,"ShowLine"):
-            if vobj.ShowLine:
-                self.lineswitch2.whichChild = -3
-                self.lineswitch3.whichChild = -3
-            else:
-                self.lineswitch2.whichChild = -1
-                self.lineswitch3.whichChild = -1
-        else:
-            self.updateData(vobj.Object,"Start")
-            
-    def remove_dim_arrows(self):
-        # remove existing nodes
-        self.node.removeChild(self.marks)
-        self.node3d.removeChild(self.marks)
-
-    def draw_dim_arrows(self, vobj):
-        from pivy import coin
-
-        if not hasattr(vobj,"ArrowType"):
-            return
-
-        if self.p3.x < self.p2.x:
-            inv = False
-        else:
-            inv = True
-
-        # set scale
-        symbol = arrowtypes.index(vobj.ArrowType)
-        s = vobj.ArrowSize.Value * vobj.ScaleMultiplier
-        self.trans1.scaleFactor.setValue((s,s,s))
-        self.trans2.scaleFactor.setValue((s,s,s))
-
-
-        # set new nodes
-        self.marks = coin.SoSeparator()
-        self.marks.addChild(self.color)
-        s1 = coin.SoSeparator()
-        if symbol == "Circle":
-            s1.addChild(self.coord1)
-        else:
-            s1.addChild(self.trans1)
-        s1.addChild(dimSymbol(symbol,invert=not(inv)))
-        self.marks.addChild(s1)
-        s2 = coin.SoSeparator()
-        if symbol == "Circle":
-            s2.addChild(self.coord2)
-        else:
-            s2.addChild(self.trans2)
-        s2.addChild(dimSymbol(symbol,invert=inv))
-        self.marks.addChild(s2)
-        self.node.insertChild(self.marks,2)
-        self.node3d.insertChild(self.marks,2)
-
-    def remove_dim_overshoot(self):
-        self.node.removeChild(self.marksDimOvershoot)
-        self.node3d.removeChild(self.marksDimOvershoot)
-
-    
-    def draw_dim_overshoot(self, vobj):
-        from pivy import coin
-
-        # set scale
-        s = vobj.DimOvershoot.Value * vobj.ScaleMultiplier
-        self.transDimOvershoot1.scaleFactor.setValue((s,s,s))
-        self.transDimOvershoot2.scaleFactor.setValue((s,s,s))
-
-        # remove existing nodes
-
-        # set new nodes
-        self.marksDimOvershoot = coin.SoSeparator()
-        if vobj.DimOvershoot.Value:
-            self.marksDimOvershoot.addChild(self.color)
-            s1 = coin.SoSeparator()
-            s1.addChild(self.transDimOvershoot1)
-            s1.addChild(dimDash((-1,0,0),(0,0,0)))
-            self.marksDimOvershoot.addChild(s1)
-            s2 = coin.SoSeparator()
-            s2.addChild(self.transDimOvershoot2)
-            s2.addChild(dimDash((0,0,0),(1,0,0)))
-            self.marksDimOvershoot.addChild(s2)
-        self.node.insertChild(self.marksDimOvershoot,2)
-        self.node3d.insertChild(self.marksDimOvershoot,2)
-
-    
-    def remove_ext_overshoot(self):
-        self.node.removeChild(self.marksExtOvershoot)
-        self.node3d.removeChild(self.marksExtOvershoot)
-
-    
-    def draw_ext_overshoot(self, vobj):
-        from pivy import coin
-
-        # set scale
-        s = vobj.ExtOvershoot.Value * vobj.ScaleMultiplier
-        self.transExtOvershoot1.scaleFactor.setValue((s,s,s))
-        self.transExtOvershoot2.scaleFactor.setValue((s,s,s))
-
-        # set new nodes
-        self.marksExtOvershoot = coin.SoSeparator()
-        if vobj.ExtOvershoot.Value:
-            self.marksExtOvershoot.addChild(self.color)
-            s1 = coin.SoSeparator()
-            s1.addChild(self.transExtOvershoot1)
-            s1.addChild(dimDash((0,0,0),(-1,0,0)))
-            self.marksExtOvershoot.addChild(s1)
-            s2 = coin.SoSeparator()
-            s2.addChild(self.transExtOvershoot2)
-            s2.addChild(dimDash((0,0,0),(-1,0,0)))
-            self.marksExtOvershoot.addChild(s2)
-        self.node.insertChild(self.marksExtOvershoot,2)
-        self.node3d.insertChild(self.marksExtOvershoot,2)
-
-
-    def doubleClicked(self,vobj):
-        self.setEdit(vobj)
-
-    def getDisplayModes(self,vobj):
-        return ["2D","3D"]
-
-    def getDefaultDisplayMode(self):
-        if hasattr(self,"defaultmode"):
-            return self.defaultmode
-        else:
-            return ["2D","3D"][getParam("dimstyle",0)]
-
-    def setDisplayMode(self,mode):
-        return mode
-
-    def is_linked_to_circle(self):
-        import DraftGeomUtils
-        _obj = self.Object
-        if _obj.LinkedGeometry and len(_obj.LinkedGeometry) == 1:
-            lobj = _obj.LinkedGeometry[0][0]
-            lsub = _obj.LinkedGeometry[0][1]
-            if len(lsub) == 1 and "Edge" in lsub[0]:
-                n = int(lsub[0][4:]) - 1
-                edge = lobj.Shape.Edges[n]
-                if DraftGeomUtils.geomType(edge) == "Circle":
-                    return True
-        return False
-
-    def getIcon(self):
-        if self.is_linked_to_circle():
-            return ":/icons/Draft_DimensionRadius.svg"
-        return ":/icons/Draft_Dimension_Tree.svg"
-
-    def __getstate__(self):
-        return self.Object.ViewObject.DisplayMode
-
-    def __setstate__(self,state):
-        if state:
-            self.defaultmode = state
-            self.setDisplayMode(state)
 
 class _AngularDimension(_DraftObject):
     """The Draft AngularDimension object"""
